requirements.txt
streamlit
Pillow
deepface
tensorflow==2.16.1
tf-keras